+++ lib/pages/v2_list_request_client/list_request_client_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:taxi_segurito_app/models/client_request.dart';
-import 'package:taxi_segurito_app/pages/list_request_client/request_list_functionality.dart';
-import 'package:taxi_segurito_app/pages/request_client_info_estimates/view_request_map.dart';
-import 'package:taxi_segurito_app/pages/list_request_client/widgets/request_list.dart';
-import 'package:taxi_segurito_app/pages/list_request_client/widgets/request_list_item.dart';
+import 'package:taxi_segurito_app/pages/v2_list_request_client/list_request_client_functionality.dart';
+import 'package:taxi_segurito_app/pages/v2_list_request_client/widgets/request_list.dart';
+import 'package:taxi_segurito_app/pages/v2_list_request_client/widgets/request_list_item.dart';
+import 'package:taxi_segurito_app/pages/v2_request_client_info_estimates/view_request_info_page.dart';
 
 class ListRequestClient extends StatefulWidget {
   ListRequestClient({Key? key}) : super(key: key);
