+++ lib/pages/v2_list_request_client/widgets/request_list_item.dart
@@ -1,32 +1,24 @@
 import 'package:flutter/material.dart';
-import 'package:geolocator/geolocator.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:taxi_segurito_app/components/buttons/CustomButtonWithLinearBorder.dart';
-import 'package:taxi_segurito_app/models/driver_request.dart';
-import 'package:taxi_segurito_app/models/company.dart';
-import 'package:taxi_segurito_app/pages/company_screen/company_edit_screen.dart';
-//import 'package:geocoding/geocoding.dart';
-import 'package:taxi_segurito_app/pages/list_request_driver/list_request_driver_functionality.dart';
-import 'package:taxi_segurito_app/pages/list_request_driver/widgets/request_list_driver_functionality.dart';
+import 'package:taxi_segurito_app/models/client_request.dart';
+import 'package:taxi_segurito_app/pages/v2_list_request_client/widgets/request_list_item_functionality.dart';
 
-class RequestListItemDriver extends StatefulWidget {
-  void Function(DriverRequest driverRequest) callbackRequest;
-  DriverRequest? driverRequest;
-  RequestListItemDriver(
-      {Key? key, this.driverRequest, required this.callbackRequest})
+class RequestListItem extends StatefulWidget {
+  void Function(ClienRequest clienRequest) callbackRequest;
+  ClienRequest? clientRequest;
+  RequestListItem({Key? key, this.clientRequest, required this.callbackRequest})
       : super(key: key);
 
   @override
   _RequestListItemState createState() => _RequestListItemState();
 }
 
-class _RequestListItemState extends State<RequestListItemDriver> {
+class _RequestListItemState extends State<RequestListItem> {
   Color colorMain = Color.fromRGBO(255, 193, 7, 1);
   Color colorMainDanger = Color.fromRGBO(242, 78, 30, 1);
   Color colorMainNull = Color.fromRGBO(153, 153, 153, 1);
-  //ListRequestDriverFunctionality listRequestDriverFunctionality = new ListRequestDriverFunctionality();
-  //RequestListItemFunctionality requestListItemFunctionality = new RequestListItemFunctionality();
-  
+  RequestListItemFunctionality requestListItemFunctionality =
+      new RequestListItemFunctionality();
+
   @override
   Widget build(BuildContext context) {
     Image imagedefault = new Image.asset(
@@ -67,14 +59,21 @@ class _RequestListItemState extends State<RequestListItemDriver> {
         children: [
           boxData(
             Text(
-              'Precio: ' + widget.driverRequest!.estimacion.toString(),
+              'Distancia: ' +
+                  requestListItemFunctionality
+                      .getDistance(widget.clientRequest!),
             ),
           ),
+          boxData(requestListItemFunctionality.getNameDirectionAddress(
+              'De: ',
+              widget.clientRequest!.latitudOrigen,
+              widget.clientRequest!.longitudOrigen)),
+          boxData(requestListItemFunctionality.getNameDirectionAddress(
+              'A: ',
+              widget.clientRequest!.latitudDestino,
+              widget.clientRequest!.longitudDestino)),
           boxData(Text(
-            'Distancia: '+ widget.driverRequest!.distancia.toString() + ' Km',
-          )),
-          boxData(Text(
-            "Placa: " + widget.driverRequest!.placa.toString(),
+            "Pasajeros: " + widget.clientRequest!.numeroPasageros.toString(),
           )),
         ],
       ),
@@ -124,7 +123,7 @@ class _RequestListItemState extends State<RequestListItemDriver> {
       child: Material(
         child: InkWell(
           onTap: () {
-            //widget.callbackRequest(widget.driverRequest!);
+            widget.callbackRequest(widget.clientRequest!);
           },
           child: Container(
             margin: new EdgeInsets.only(
