+++ lib/pages/v2_request_client_info_estimates/view_request_info_page.dart
@@ -6,9 +6,9 @@ import 'package:location/location.dart';
 import 'package:taxi_segurito_app/components/buttons/CustomButton.dart';
 import 'package:taxi_segurito_app/components/inputs/CustomTextField.dart';
 import 'package:taxi_segurito_app/models/taxi_request.dart';
-import 'package:taxi_segurito_app/pages/request_client_info_estimates/view_request_map_functionality.dart';
+import 'package:taxi_segurito_app/strategis/convert_distance.dart';
+import 'package:taxi_segurito_app/pages/v2_request_client_info_estimates/view_request_info_functionality.dart';
 import 'package:taxi_segurito_app/validators/TextFieldValidators.dart';
-import 'package:taxi_segurito_app/pages/list_request_client/convert_distance.dart';
 
 import '../../models/client_request.dart';
