+++ lib/strategis/convert_distance.dart
@@ -3,7 +3,7 @@ import 'package:geocoding/geocoding.dart';
 import 'package:geolocator/geolocator.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 
-import '../../../models/client_request.dart';
+import '../../models/client_request.dart';
 
 class ConvertDistance {
   String getDistance(ClienRequest clientRequest) {
